Refactored Models package & improved Location Update
