+++ app/src/main/java/org/wit/placemark/main/MainApp.kt
@@ -3,9 +3,8 @@ package org.wit.placemark.main
 import android.app.Application
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
-import org.wit.placemark.models.HillfortJSONStore
-import org.wit.placemark.models.HillfortMemStore
 import org.wit.placemark.models.HillfortStore
+import org.wit.placemark.models.json.HillfortJSONStore
 
 class MainApp : Application(), AnkoLogger {
