+++ app/src/main/java/org/wit/placemark/views/hillfortlist/HillfortListPresenter.kt
@@ -22,4 +22,5 @@ class HillfortListPresenter(view: BaseView) : BasePresenter(view) {
     fun loadHillforts() {
         view?.showHillforts(app.hillforts.findAll())
     }
+
 }
