+++ app/src/main/java/org/wit/placemark/views/location/EditLocationPresenter.kt
@@ -27,6 +27,7 @@ class EditLocationPresenter(view: BaseView) : BasePresenter(view) {
             .position(loc)
         map.addMarker(options)
         map.moveCamera(CameraUpdateFactory.newLatLngZoom(loc, location.zoom))
+        view?.showLocation(loc.latitude, loc.longitude);
     }
 
     fun doUpdateLocation(lat: Double, lng: Double) {
