+++ app/src/main/java/org/wit/placemark/views/location/EditLocationView.kt
@@ -82,4 +82,9 @@ class EditLocationView : BaseView(), GoogleMap.OnMarkerDragListener, GoogleMap.O
     super.onSaveInstanceState(outState)
     mapView.onSaveInstanceState(outState)
   }
+
+  override fun showLocation(latitude : Double, longitude : Double) {
+    lat.setText("%.6f".format(latitude))
+    lng.setText("%.6f".format(longitude))
+  }
 }
